+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Implemented Advanced Search Query to URL
 * `Source` column hidden on flaw lists
 * `Source` field renamed to `CVE Source` on flaw form and advanced search
+* Added Total count for IndexView and AdvancedSearch Page
 
 ### Fixed
 * Fixed overlap on edit buttons at References and Acknowledgements items